+++ recipes/default.rb
@@ -2,7 +2,28 @@
 # Cookbook Name:: prometheus
 # Recipe:: default
 #
-# Copyright (C) 2015 YOUR_NAME
+# Author: Ray Rodriguez <rayrod2030@gmail.com>
 #
-# All rights reserved - Do Not Redistribute
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+include_recipe 'runit::default'
+
+include_recipe "prometheus::#{node['prometheus']['install_method']}"
+
+# rubocop:disable Style/HashSyntax
+service 'prometheus' do
+  supports :restart => true
+  action :start
+end
+# rubocop:enable Style/HashSyntax