+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-# 0.1.0
+# Change Log
+All notable changes to this project will be documented in this file.
 
-Initial release of prometheus
+## 0.1.0 - 2015-02-28
+### Changed
+- Initial release of prometheus cookbook